nd: added initial api call in app_nd.js
