+++ package.json
@@ -5,9 +5,12 @@
   "main": "app.js",
   "dependencies": {
     "express": "^4.16.2",
+    "express-handlebars": "^3.0.0",
     "handlebars": "^4.0.11",
     "jquery": "^3.3.1",
-    "request": "^2.83.0"
+    "md5": "^2.2.1",
+    "request": "^2.83.0",
+    "request-promise": "^4.2.2"
   },
   "devDependencies": {
     "morgan": "^1.9.0"
